[Fix] fix course service after test

+++ Edulingual.Service/Request/Course/UpdateCourseRequest.cs
@@ -8,7 +8,7 @@ public class UpdateCourseRequest
     public string? Title { get; set; } = string.Empty;
     public string? Description { get; set; } = string.Empty;
     public string? Duration { get; set; } = string.Empty;
-    public decimal? Fee { get; set; }
+    public double? Fee { get; set; }
     public Guid? CourseAreaId { get; set; }
     public Guid? CourseLanguageId { get; set; }
     public Guid? CourseCategoryId { get; set; }
